Fix some MSVC warnings
And update the CMakeLists.txt VERSION to 1.0.0

+++ examples/01_simple_struct.cpp
@@ -41,7 +41,7 @@ struct TypeHandler<ModuleList>
       return JS::Error::ExpectedArrayStart;
 
     context.nextToken();
-    for (size_t i = 0; i < ModuleList::ReservedSize; i++)
+    for (int i = 0; i < int(ModuleList::ReservedSize); i++)
     {
       if (context.error != JS::Error::NoError)
         return context.error;
